+++ changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - upgrade: update spring-boot from v2.5.6 to v2.7.1
 - upgrade: update electrumx testcontainer from v1.15.0 to v1.16.0
+- upgrade: update bitcoin testcontainer from v22 to v23
 - upgrade: update bitcoin testcontainer from v0.21.1 to v22
 - upgrade: update bitcoin-kmp from v0.8.1 to v0.8.2
 - upgrade: update bitcoinj from v0.15.10 to v0.16.1